Move default header to subfolder
