+++ example/ExtremeScaling/bunny.cxx
@@ -26,7 +26,7 @@
 #include <t8_cmesh_tetgen.h>
 #include <t8_cmesh_vtk.h>
 #include <t8_forest.h>
-#include <t8_schemes/t8_default_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>
 #include <p8est.h>
 #include <p8est_connectivity.h>
 #include <p8est_tets_hexes.h>
