+++ example/basic/t8_basic.cxx
@@ -22,7 +22,7 @@
 
 #include <sc_options.h>
 #include <sc_refcount.h>
-#include <t8_schemes/t8_default_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>
 #include <t8_forest.h>
 #include <t8_cmesh_vtk.h>
 #include <t8_cmesh/t8_cmesh_examples.h>
