+++ example/cmesh/t8_cmesh_load_save.cxx
@@ -25,7 +25,7 @@
 #include <t8_cmesh_vtk.h>
 #include <t8_cmesh_readmshfile.h>
 #include <t8_cmesh/t8_cmesh_examples.h>
-#include <t8_schemes/t8_default_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>
 
 /* TODO: rename this file to t8_something */
