+++ example/cmesh/t8_cmesh_partition.cxx
@@ -29,7 +29,7 @@
 #include <p4est_connectivity.h>
 #include <p8est_connectivity.h>
 #include <sc_shmem.h>
-#include <t8_schemes/t8_default_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>
 
 /* TODO: rename this file to t8_something */
