+++ example/ghost/t8_test_face_iterate.cxx
@@ -25,7 +25,7 @@
 #include <sc_refcount.h>
 #include <t8_eclass.h>
 #include <t8_element_cxx.hxx>
-#include <t8_schemes/t8_default_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>
 #include <t8_forest.h>
 #include <t8_cmesh.h>
 #include <t8_cmesh_readmshfile.h>
