+++ example/gmsh/t8_load_and_refine_square_w_hole.cxx
@@ -32,7 +32,7 @@
 #include <t8_cmesh_readmshfile.h>
 #include <t8_forest.h>
 #include <t8_forest_vtk.h>
-#include <t8_schemes/t8_default_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>
 
 /* Simple 3 dimensional vector product */
 static double
