+++ example/timings/time_forest_partition.cxx
@@ -31,7 +31,7 @@
 #include <t8_cmesh/t8_cmesh_partition.h>
 #include <t8_cmesh_readmshfile.h>
 #include <t8_forest.h>
-#include <t8_schemes/t8_default_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>
 #include <example/common/t8_example_common.h>
 
 /* This is the user defined data used to define the
