+++ example/triangle/t8_read_triangle_file.cxx
@@ -25,7 +25,7 @@
 #include <t8_cmesh.h>
 #include <t8_cmesh_triangle.h>
 #include <t8_cmesh_vtk.h>
-#include <t8_schemes/t8_default_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>
 
 void
 t8_read_triangle_file_build_cmesh (const char *prefix, int do_dup,
