+++ example/tutorials/t8_step4_partition_balance_ghost.cxx
@@ -89,7 +89,7 @@
 #include <t8_cmesh.h>           /* cmesh definition and basic interface. */
 #include <t8_cmesh/t8_cmesh_examples.h> /* A collection of exemplary cmeshes */
 #include <t8_forest.h>          /* forest definition and basic interface. */
-#include <t8_schemes/t8_default_cxx.hxx>        /* default refinement scheme. */
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>        /* default refinement scheme. */
 #include <example/tutorials/t8_step3.h>
 
 T8_EXTERN_C_BEGIN ();
