+++ example/tutorials/t8_step5_element_data.cxx
@@ -31,7 +31,7 @@
 #include <t8_cmesh.h>           /* cmesh definition and basic interface. */
 #include <t8_cmesh/t8_cmesh_examples.h> /* A collection of exemplary cmeshes */
 #include <t8_forest.h>          /* forest definition and basic interface. */
-#include <t8_schemes/t8_default_cxx.hxx>        /* default refinement scheme. */
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>        /* default refinement scheme. */
 #include <t8_forest_vtk.h>      /* Additional vtk functions to output arbitrary user data. */
 #include <example/tutorials/t8_step3.h>
