+++ test/t8_gtest_nca.cxx
@@ -27,7 +27,7 @@ along with t8code; if not, write to the Free Software Foundation, Inc.,
 
 #include <gtest/gtest.h>
 #include <t8_eclass.h>
-#include <t8_schemes/t8_default_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>
 
 /* *INDENT-OFF* */
 class nca:public testing::TestWithParam < t8_eclass > {
