+++ test/t8_test_element_count_leafs.cxx
@@ -20,7 +20,7 @@
   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 */
 
-#include <t8_schemes/t8_default_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>
 
 /*
  * In this file we test whether the t8_element_count_leafs{_from_root}
