+++ test/t8_test_element_general_function.cxx
@@ -21,7 +21,7 @@
 */
 
 #include <t8.h>
-#include <t8_schemes/t8_default_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>
 #include <t8_schemes/t8_default/t8_default_tri/t8_dtri.h>
 #include <t8_schemes/t8_default/t8_default_tet/t8_dtet.h>
 #include <t8_schemes/t8_default/t8_default_prism/t8_dprism.h>
