+++ test/t8_test_forest_commit.cxx
@@ -23,7 +23,7 @@
 #include <t8_eclass.h>
 #include <t8_cmesh.h>
 #include <t8_forest.h>
-#include <t8_schemes/t8_default_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>
 #include <t8_forest/t8_forest_partition.h>
 #include <t8_forest/t8_forest_private.h>
 #include "t8_cmesh/t8_cmesh_testcases.h"
