+++ test/t8_test_point_inside.cxx
@@ -28,7 +28,7 @@
 #include <t8_cmesh/t8_cmesh_examples.h>
 #include <t8_forest.h>
 #include <t8_forest/t8_forest_iterate.h>
-#include <t8_schemes/t8_default_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>
 #include <t8_element_cxx.hxx>
 #include <t8_cmesh/t8_cmesh_geometry.h>
 #include <t8_geometry/t8_geometry_implementations/t8_geometry_linear.hxx>
