+++ test/t8_test_user_data.cxx
@@ -24,7 +24,7 @@
  * t8_forest_set/get_user_function functions.
  */
 
-#include <t8_schemes/t8_default_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_cxx.hxx>
 #include <t8_cmesh.h>
 #include <t8_cmesh/t8_cmesh_examples.h>
 #include <t8_forest.h>
